requirements.txt
django
djangorestframework
djangorestframework-jwt
psycopg2
dj-database-url
gunicorn
django-cors-headers
